# api/api/openapi.yaml
openapi: 3.0.3
info:
  title: UpcycleConnect API
  version: 1.0.0
  description: API REST pour UpcycleConnect (admin, pro, particulier)
servers:
  - url: http://localhost:8080
paths:
...
  /conteneurs/{id}:
    put:
      summary: Mettre à jour conteneur (admin/staff)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Mis à jour }
    delete:
      summary: Supprimer conteneur (admin/staff)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Supprimé }
...
  /conseils:
    get:
      tags: [Conseils]
      summary: Lister les conseils/news publiés
      description: Retourne uniquement les contenus avec is_active=true (catalogue public).
      security: []
      responses:
        '200':
          description: Liste des conseils publiés
        '500':
          description: Erreur serveur
    post:
      tags: [Conseils]
      summary: Créer un conseil/news
      description: |
        Staff/admin peut publier directement (is_active=true).
        Salarié (role_id=5) : is_active forcé à false côté serveur (brouillon), id_auteur = utilisateur connecté.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConseilInput'
      responses:
        '201':
          description: Conseil créé
        '400':
          description: Champs invalides (titre/contenu requis)
        '401':
          description: Non authentifié
        '403':
          description: Rôle non autorisé
        '500':
          description: Erreur serveur
  /conseils/{id}:
    get:
      tags: [Conseils]
      summary: Lire un conseil/news publié
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK }
        '404': { description: Non trouvé ou non publié }
        '500': { description: Erreur serveur }
    put:
      tags: [Conseils]
      summary: Mettre à jour un conseil/news
      description: |
        Salarié : uniquement ses brouillons (is_active reste false).
        Staff/admin : peut publier (is_active=true) ou repasser en brouillon.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConseilInput'
      responses:
        '200': { description: Mis à jour }
        '400': { description: Données invalides }
        '401': { description: Non authentifié }
        '403': { description: Interdit (auteur, publication, etc.) }
        '404': { description: Non trouvé }
        '500': { description: Erreur serveur }
    delete:
      tags: [Conseils]
      summary: Supprimer un conseil/news
      description: Staff/admin supprime tout contenu. Salarié supprime uniquement ses brouillons.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Supprimé }
        '401': { description: Non authentifié }
        '403': { description: Interdit }
        '404': { description: Non trouvé }
        '500': { description: Erreur serveur }
  /admin/conseils:
    get:
      tags: [Conseils, Admin]
      summary: Liste admin de tous les conseils/news
      description: |
        Brouillons et publiés, avec auteur. Réservé admin (role 1) et staff (role 2).
        Filtres query : status (draft|published), author (id_auteur), q, date_from, date_to, page, per_page.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema: { type: string, enum: [draft, published, brouillon, publie, active] }
        - in: query
          name: author
          schema: { type: integer }
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: date_from
          schema: { type: string, format: date }
        - in: query
          name: date_to
          schema: { type: string, format: date }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, default: 25 }
      responses:
        '200':
          description: Liste paginée (items, total, page, per_page)
        '401': { description: Non authentifié }
        '403': { description: Réservé admin/staff }
        '500': { description: Erreur serveur }
...
  /me/conseils:
    get:
      tags: [Conseils]
      summary: Mes conseils/news (brouillons et publiés de l'auteur)
      description: Salarié ou tout auteur connecté — uniquement ses propres contenus.
      security:
        - bearerAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Non authentifié }
        '500': { description: Erreur serveur }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via POST /login
  schemas:
    ConseilInput:
      type: object
      required: [titre, contenu]
      properties:
        titre: { type: string, maxLength: 150 }
        contenu: { type: string }
        categorie: { type: string, maxLength: 100 }
        image_url: { type: string, maxLength: 255 }
        is_active:
          type: boolean
          description: Ignoré pour salarié (forcé false). Utilisé par staff/admin pour publier.
...